+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * expose schema types which define `parameters` with new `ParameterOrRef` type
 * Adds License object
 * Adds Contact object
+* Derives Default for all structs
 
 # 0.1.5